+++ wn/__init__.py
@@ -31,10 +31,17 @@ __all__ = (
     'ILI',
     'Error',
     'DatabaseError',
+    'ConfigurationError',
     'WnWarning',
 )
 
-from wn._exceptions import Error, DatabaseError, WnWarning
+from wn._exceptions import (
+    Error,
+    DatabaseError,
+    ConfigurationError,
+    ProjectError,
+    WnWarning,
+)
 from wn._config import config  # noqa: F401
 from wn._add import add, remove
 from wn._export import export